Move logic from model to manager
- remove docstrings because the author of the project doesn't like them :(
- override default managers
