+++ comment/managers/__init__.py
@@ -1,3 +1,4 @@
 # flake8: noqa
 from comment.managers.comments import *
 from comment.managers.reactions import *
+from comment.managers.flags import *
