+++ comment/models/comments.py
@@ -1,4 +1,3 @@
-from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.contenttypes.fields import GenericForeignKey
@@ -18,8 +17,6 @@ class Comment(models.Model):
     edited = models.DateTimeField(auto_now=True)
 
     objects = CommentManager()
-
-    ALLOWED_FLAGS = getattr(settings, 'COMMENT_FLAGS_ALLOWED', 10)
 
     class Meta:
         ordering = ['-posted', ]
@@ -55,3 +52,7 @@ class Comment(models.Model):
     @property
     def dislikes(self):
         return self._get_reaction_count('dislikes')
+
+    @property
+    def is_flagged(self):
+        return self.flag.count > self.objects.ALLOWED_FLAGS
